Moved around some components and added an additional component for the SearchQuotes to be able to display the results from the API
