+++ src/components/App/App.jsx
@@ -21,6 +21,7 @@ import RegisterPage from '../RegisterPage/RegisterPage';
 
 import './App.css';
 import Quotes from '../Quotes/Quotes';
+import SearchQuotes from '../Quotes/SearchQuotes'
 import RandomQuotes from '../Quotes/RandomQuotes';
 
 function App() {
@@ -73,7 +74,7 @@ function App() {
           </ProtectedRoute>
 
           <ProtectedRoute exact path="/quotes">
-            <Quotes />
+            <SearchQuotes />
           </ProtectedRoute>
 
           <Route
