+++ src/redux/sagas/quote.saga.js
@@ -3,7 +3,8 @@ import { takeEvery, put} from 'redux-saga/effects';
 
 function* getQuote(action){
     try{
-        const quotes = yield axios.get(`/api/search/quotes`, action.payload);
+        console.log('quote saga console.log', action.payload);
+        const quotes = yield axios.post(`/api/search/quotes&query=${action.payload}`);
         console.log('Checking quotes.data', quotes.data)
         yield put({ type: 'SET_QUOTES', payload: quotes.data });
